+++ packages/core/styles/src/gradients/index.ts
@@ -2,3 +2,8 @@ import * as gradients from "./gradients";
 import { createLinearGradient, createBorderGradient } from "./gradientBuilder";
 
 export { gradients, createLinearGradient, createBorderGradient };
+export {
+  aiGenerativeInteractive,
+  aiGenerativeHover,
+  aiGenerativeDisabled,
+} from "./gradients";