+++ packages/react/src/atomic/Text/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Text is a basic component that allows us to write blocks of text and give it formatting to use within other components, sections and pages of our application or website.
 
+## 2025-09-26 `6.5.0`
+
+#### 🎉 New features
+
+- Added `ai-generative` appearance to `Text` component. ([#363](https://github.com/TiendaNube/nimbus-design-system/pull/363) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-03-25 `6.4.0`
 
 ### 🎉 New features